♻️ refactor: çommands 분리 + normalWorking이름 변경 racingcar.spec.js의 commands들을 support폴더 내의 commands.js에서 정의 + constatns.js에 사용하는 상수들 정의 + normalWorking을 자동차 이름 입력 , 시도횟수 입력받는 이벤트로 분리함
