+++ cypress/integration/racingcar.spec.js
@@ -1,53 +1,5 @@
+import {baseUrl} from '../support/constants.js';
 describe('구현 결과가 요구사항과 일치해야 한다.', () => {
-  const baseUrl = "../../index.html";
-  const SELECTOR = {
-    CAR_NAME_INPUT: "#car-name-input",
-    CAR_NAME_BUTTON: '#car-name-button',
-    RACE_COUNT_INPUT: '#race-count-input',
-    RACE_COUNT_BUTTON: '#race-count-button',
-    RESULT_TEXT: '.result-text',
-    RESTART_BUTTON: '.restart-button',
-  };
-
-  before(() => {
-    Cypress.Commands.add('normalWorking', (carNames, racingCount) => {
-      cy.get(SELECTOR.CAR_NAME_INPUT).type(carNames);
-      cy.get(SELECTOR.CAR_NAME_BUTTON).click();
-  
-      cy.get(SELECTOR.RACE_COUNT_INPUT).type(racingCount);
-      cy.get(SELECTOR.RACE_COUNT_BUTTON).click();
-    });
-
-    Cypress.Commands.add('showCarNameAlert', (invalidInput) => {
-      const alertStub = cy.stub();
-      cy.on("window:alert", alertStub);
-    
-      cy.get(SELECTOR.CAR_NAME_INPUT).type(invalidInput);
-    
-      cy.get(SELECTOR.CAR_NAME_BUTTON)
-        .click()
-        .then(() => {
-          expect(alertStub).to.be.called;
-      });
-    });
-
-    Cypress.Commands.add('showRaceCountAlert', (carNameInput, invalidInput) => {
-      const alertStub = cy.stub();
-      cy.get(SELECTOR.CAR_NAME_INPUT).type(carNameInput);
-      cy.get(SELECTOR.CAR_NAME_BUTTON).click();
-
-      cy.on("window:alert", alertStub);
-
-      cy.get(SELECTOR.RACE_COUNT_INPUT).type(invalidInput);
-
-      cy.get(SELECTOR.RACE_COUNT_BUTTON)
-        .click()
-        .then(() => {
-          expect(alertStub).to.be.called;
-      });
-    });
-
-  });
 
   beforeEach(() => {
     cy.visit(baseUrl);
@@ -55,8 +7,9 @@ describe('구현 결과가 요구사항과 일치해야 한다.', () => {
 
   /* 우승자 확인 */
   it("게임을 완료하고 우승자를 확인할 수 있어야 한다.", () => {
+    cy.carNamePositiveInputEvent("tt,sally");
+    cy.raceCountPositiveInputEvent(1);
     
-    cy.normalWorking("tt,sally", 1);
     cy.get('.result-text').should((result) => {
       const text = result.text();
       expect(text).to.include('최종 우승자')
@@ -79,9 +32,10 @@ describe('구현 결과가 요구사항과 일치해야 한다.', () => {
 
   /* 다시 시작 */
   it("다시 시작하기 버튼을 눌렀을 때에 race-count-input-container 요소가 display none이어야 한다", () => {
-    cy.normalWorking("tt,sally", 1);
+    cy.carNamePositiveInputEvent("tt,sally");
+    cy.raceCountPositiveInputEvent(1);
 
-    cy.get('.restart-button').click();
+    cy.get('.restart-button').contains('다시 시작하기').click();
 
     cy.get('.race-count-input-container').should('not.have.css', 'display', 'flex');
   });
