+++ cypress/support/constants.js
@@ -1,0 +1,10 @@
+export const baseUrl = "../../index.html";
+export const SELECTOR = {
+    CAR_NAME_INPUT: "#car-name-input",
+    CAR_NAME_BUTTON: '#car-name-button',
+    RACE_COUNT_INPUT: '#race-count-input',
+    RACE_COUNT_BUTTON: '#race-count-button',
+    RESULT_TEXT: '.result-text',
+    RESTART_BUTTON: '.restart-button',
+};
+
