+++ 리뷰개선사항들.md
@@ -1,8 +1,8 @@
 ## racingcar.spec.js 에서의 개선사항 
 
-- [ ] 전역에서 사용하는 Custom Commands들을 support폴더에서 정의하기 
-- [ ] https://docs.cypress.io/guides/references/best-practices#Selecting-Elements 해당문서 참조하여 app.spec.js 리팩토링 하기
-- [ ] normalWorking을 이름을 입력하는 부분과 경기 횟수를 입력하는 부분을 나눠보기
+- [x] 전역에서 사용하는 Custom Commands들을 support폴더에서 정의하기 
+- [x] https://docs.cypress.io/guides/references/best-practices#Selecting-Elements 해당문서 참조하여 app.spec.js 리팩토링 하기
+- [x] normalWorking을 이름을 입력하는 부분과 경기 횟수를 입력하는 부분을 나눠보기
 
 ## racingCarGame.js 에서의 개선사항
